rm: remove old mock state
